Ädd routes handling to the controllers
