+++ src/controllers/ClientController.ts
@@ -1,8 +1,16 @@
-import { Request, Response } from 'express'
+import { Request, Response, Router } from 'express'
 
 import { IClientController } from '@src/interfaces'
 
 export class ClientController implements IClientController {
+  async init(route: Router): Promise<void> {
+    route.get('/clients', this.list)
+    route.post('/clients', this.create)
+    route.get('/clients/:id', this.listOne)
+    route.put('/clients/:id', this.update)
+    route.delete('/clients/:id', this.remove)
+  }
+
   async list(_: Request, response: Response): Promise<Response> {
     return response.json('Listing clients ...')
   }
