+++ src/controllers/ClientStatusController.ts
@@ -1,7 +1,15 @@
 import { IClientStatusController } from '@src/interfaces'
-import { Request, Response } from 'express'
+import { Request, Response, Router } from 'express'
 
 export class ClientStatusController implements IClientStatusController {
+  async init(route: Router): Promise<void> {
+    route.get('/clients-status', this.list)
+    route.post('/clients-status', this.create)
+    route.get('/clients-status/:id', this.listOne)
+    route.put('/clients-status/:id', this.update)
+    route.delete('/clients-status/:id', this.remove)
+  }
+
   async list(_: Request, response: Response): Promise<Response> {
     return response.json('Listing client status...')
   }
