+++ src/interfaces/IApp.ts
@@ -1,7 +1,7 @@
-export interface IApp {
-  init(): Promise<void>
-  start(): void
-  initDatabase(): Promise<void>
-  initMiddlewares(): void
-  initRoutes(): void
-}
+// export interface IApp {
+//   init(): Promise<void>
+//   start(): void
+//   initDatabase(): Promise<void>
+//   initMiddlewares(): void
+//   initRoutes(): void
+// }
